Adding explicit name to sliding window alg, code cleanup
