+++ Reconstruction/RecCalorimeter/src/components/CreateCaloClustersSlidingWindow.cpp
@@ -1,0 +1,58 @@
+#include "CreateCaloClustersSlidingWindow.h"
+
+#include <unordered_map>
+#include <utility>
+#include <boost/functional/hash.hpp>
+
+// datamodel
+#include "datamodel/CaloHitCollection.h"
+#include "datamodel/CaloClusterCollection.h"
+
+DECLARE_ALGORITHM_FACTORY(CreateCaloClustersSlidingWindow)
+
+CreateCaloClustersSlidingWindow::CreateCaloClustersSlidingWindow(const std::string& name, ISvcLocator* svcLoc)
+  : GaudiAlgorithm(name, svcLoc) {
+  declareInput("cells", m_cells,"calo/cells");
+  declareOutput("clusters", m_clusters,"calo/clusters");
+  declareProperty("buildTowersTool",m_buildTowersTool);
+  declarePrivateTool(m_buildTowersTool,"BuildCaloTowersTool");
+}
+
+CreateCaloClustersSlidingWindow::~CreateCaloClustersSlidingWindow() {}
+
+StatusCode CreateCaloClustersSlidingWindow::initialize() {
+  if (GaudiAlgorithm::initialize().isFailure()) {
+    return StatusCode::FAILURE;
+  }
+  info() << "CreateCaloClustersSlidingWindow initialized" << endmsg;
+  //Build calo towers
+  if (!m_buildTowersTool.retrieve()) {
+    error()<<"Unable to retrieve the build towers tool tool!"<<endmsg;
+    return StatusCode::FAILURE;
+  }
+  return StatusCode::SUCCESS;
+}
+
+StatusCode CreateCaloClustersSlidingWindow::execute() {
+  // Get the input collection with cells from simulation + digitisation (after calibration and with noise)
+  const fcc::CaloHitCollection* cells = m_cells.get();
+  debug() << "Input Hit collection size: " << cells->size() << endmsg;
+
+  // 1. Get calorimeter towers (calorimeter grid in eta phi, all layers merged)
+  auto caloTowers = m_buildTowersTool->buildTowers(*cells);
+
+  // 2. Find local maxima with sliding window
+
+  // 3. Calculate barycentre position
+
+  // 4. Remove duplicates
+
+  // 5. Create final clusters
+
+  return StatusCode::SUCCESS;
+}
+
+StatusCode CreateCaloClustersSlidingWindow::finalize() {
+  StatusCode sc = GaudiAlgorithm::finalize();
+  return sc;
+}
